move mocked_rdmo into dedicated pytest plugin
